chore(backend):added middlewares
