+++ src/controllers/blog.controller.js
@@ -1,10 +1,8 @@
 import Blog from '../database/model/blog.model';
 import { blogValidation } from "../validate/index";
-import Inquiry from '../database/model/inquiry.model';
-import newsLetters from '../database/model/newsletter.model';
-import { fileUpload } from "../helpers/file";
 
-//blogs
+
+import { fileUpload } from "../middleware/file";
 
 export const saveBlog = async(req, res, next) => {
     const { error } = blogValidation(req.body);
@@ -52,41 +50,13 @@ export const deleteBlogById = async(req, res) => {
     res.status(200).json({ status: "success", message: "Blog deleted" });
 }
 
-//query
-
-export const getAllInquiries = async(req, res) => {
-    const inquiries = await Inquiry.find();
-    res.status(200).json({ status: true, data: inquiries })
-}
-
-export const getInquiryById = async(req, res) => {
+export const getblogById = async(req, res) => {
     const { id } = req.params;
-    const inquiry = await Inquiry.findById(id);
-    if (!inquiry) return res.status(404).json({ status: "fail", message: "Inquiry not found" });
-    res.status(200).json({ status: true, data: inquiry });
+    const blog = await Blog.findById(id);
+    if (!blog) return res.status(404).json({ status: "fail", message: "Blog not found" });
+    res.status(200).json({ status: true, data: blog });
 }
-
-export const deleteInquiryById = async(req, res) => {
-    const { id } = req.params;
-    const inquiry = await Inquiry.findById(id);
-    if (!inquiry) return res.status(404).json({ status: "fail", message: "Inquiry not found" });
-    await Inquiry.findByIdAndDelete(id);
-    res.status(200).json({ status: "success", message: "Inquiry deleted", data: "Null" });
-}
-
-
-//newsletter
-
-
-export const getAllSubscribers = async(req, res) => {
-    const news = await newsLetters.find();
-    res.status(200).json({ status: true, data: news })
-}
-
-export const deleteNewsById = async(req, res) => {
-    const { id } = req.params;
-    const news = await newsLetters.findById(id);
-    if (!news) return res.status(404).json({ status: "fail", message: "email not found" });
-    await newsLetters.findByIdAndDelete(id);
-    res.status(200).json({ status: "success", message: "email deleted", data: "Null" });
+export const getAllBlogs = async(req, res) => {
+    const blogs = await Blog.find();
+    res.status(200).json({ status: true, data: blogs })
 }
