+++ src/index.js
@@ -1,8 +1,10 @@
 import "./database";
 import express from 'express';
-import userRoutes from './routes/users.route';
-import adminRoutes from './routes/admin.route';
+import inquiryRoutes from './routes/inquiry.route';
+import blogRoutes from './routes/blog.route';
 import authRoutes from './routes/authentication.route';
+import commentRoutes from './routes/comment.route';
+import newsletterRoutes from './routes/newsletter.route';
 
 const server = express();
 
@@ -13,9 +15,11 @@ server.get('/', (req, res) => {
 
 server.use(express.json());
 
-server.use('/api/v1/users', userRoutes);
-server.use('/api/v1/admin/', adminRoutes);
+server.use('/api/v1/inquiry', inquiryRoutes);
+server.use('/api/v1/blog', blogRoutes);
 server.use('/api/v1/authentication', authRoutes);
+server.use('/api/v1/newsletter', newsletterRoutes);
+server.use('/api/v1/comment', commentRoutes);
 
 
 const port = process.env.PORT;
