+++ src/middleware/hash-password.js
