+++ src/routes/blog.route.js
@@ -1,6 +1,6 @@
 import express from 'express';
-import { deleteBlogById, saveBlog, updateBlog, getAllInquiries, getInquiryById, deleteInquiryById, deleteNewsById, getAllSubscribers } from '../controllers/admin.controller';
-import { checkAdmin } from "../helpers/check";
+import { deleteBlogById, saveBlog, updateBlog, getAllBlogs, getblogById } from '../controllers/blog.controller';
+import { checkAdmin } from "../middleware/check";
 const router = express.Router();
 //image
 import multer from "multer";
@@ -17,10 +17,7 @@ const uploads = multer({ storage, fileFilter });
 router.post('/', checkAdmin, uploads.single("image"), saveBlog);
 router.put('/:id', checkAdmin, updateBlog);
 router.delete('/:id', checkAdmin, deleteBlogById);
-router.get('/query', checkAdmin, getAllInquiries);
-router.get('/query/:id', checkAdmin, getInquiryById);
-router.delete('/query/:id', checkAdmin, deleteInquiryById);
-router.delete('/newsletter/:id', checkAdmin, deleteNewsById);
-router.get('/newsletter/id', checkAdmin, getAllSubscribers);
+router.get('/', getAllBlogs);
+router.get('/:id', getblogById);
 
 export default router;
